Elaborate to Problems instead of Core.
