+++ src/Path/Elab.hs
@@ -4,235 +4,165 @@ module Path.Elab where
 import Control.Effect
 import Control.Effect.Error
 import Control.Effect.Reader hiding (Reader(Local))
-import Control.Effect.State
 import Control.Effect.Writer
 import Control.Monad (foldM)
 import Data.Bifunctor (first)
+import Data.Foldable (foldl')
 import Data.Functor.Const
 import qualified Data.Map as Map
-import qualified Data.Set as Set
 import Data.Void
-import Path.Constraint hiding ((|-))
-import Path.Context as Context
-import Path.Core
 import Path.Error
 import Path.Module as Module
 import Path.Name
 import Path.Plicity
 import Path.Pretty
+import Path.Problem
 import Path.Scope
-import Path.Semiring
-import Path.Solver
 import Path.Span
 import Path.Stack as Stack
 import qualified Path.Surface as Surface
+import Path.Term
 import Path.Usage
 import Prelude hiding (pi)
 
 assume :: ( Carrier sig m
           , Member (Error Doc) sig
-          , Member (Reader (ModuleGraph Core Void)) sig
+          , Member (Reader Context) sig
           , Member (Reader Span) sig
           )
        => Qualified
-       -> m (Core (Name Meta) ::: Core (Name Meta))
-assume n = asks @(ModuleGraph Core Void) (fmap (weaken . unSpanned . declType) . Module.lookup n) >>= maybe (freeVariables (pure n)) (pure . (global n :::))
+       -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+assume v = asks (lookupBinding (Global v)) >>= maybe (freeVariables (pure v)) (pure . (Var (Global v) :::) . typedType)
 
 intro :: ( Carrier sig m
          , Member Naming sig
-         , Member (Reader (Context (Core (Name Meta)))) sig
-         , Member (Reader Span) sig
-         , Member (State Signature) sig
-         , Member (Writer (Set.Set (Spanned (Constraint Core (Name Meta))))) sig
+         , Member (Reader Context) sig
          )
-      => Plicit (Maybe User)
-      -> m (Core (Name Meta) ::: Core (Name Meta))
-      -> m (Core (Name Meta) ::: Core (Name Meta))
-intro (p :< _) body = do
-  _A <- exists Type
+      => m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+      -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+intro body = do
+  _A <- meta type'
   x <- fresh
-  _B <- x ::: _A |- exists Type
-  u <- x ::: _A |- goalIs _B body
-  pure (lam (p :< Local (Name x)) u ::: pi (p :< Local (Name x) ::: More :@ _A) _B)
+  _B <- ForAll x ::: _A |- meta type'
+  u <- ForAll x ::: _A |- goalIs _B body
+  pure (lam (Local x) u ::: pi (Local x ::: _A) _B)
 
 (-->) :: ( Carrier sig m
          , Member Naming sig
-         , Member (Reader (Context (Core (Name Meta)))) sig
-         , Member (Reader Span) sig
-         , Member (State Signature) sig
-         , Member (Writer (Set.Set (Spanned (Constraint Core (Name Meta))))) sig
+         , Member (Reader Context) sig
          )
-      => Plicit (Maybe User, Used (m (Core (Name Meta) ::: Core (Name Meta))))
-      -> m (Core (Name Meta) ::: Core (Name Meta))
-      -> m (Core (Name Meta) ::: Core (Name Meta))
-(p :< (_, m :@ t)) --> body = do
-  t' <- goalIs Type t
+      => m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+      -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+      -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+t --> body = do
+  t' <- goalIs type' t
   x <- fresh
-  b' <- x ::: t' |- goalIs Type body
-  pure (pi (p :< Local (Name x) ::: m :@ t') b' ::: Type)
+  b' <- ForAll x ::: t' |- goalIs type' body
+  pure (pi (Local x ::: t') b' ::: type')
 
 app :: ( Carrier sig m
        , Member Naming sig
-       , Member (Reader (Context (Core (Name Meta)))) sig
-       , Member (Reader Span) sig
-       , Member (State Signature) sig
-       , Member (Writer (Set.Set (Spanned (Constraint Core (Name Meta))))) sig
+       , Member (Reader Context) sig
        )
-    => m (Core (Name Meta) ::: Core (Name Meta))
-    -> Plicit (m (Core (Name Meta) ::: Core (Name Meta)))
-    -> m (Core (Name Meta) ::: Core (Name Meta))
-app f (p :< a) = do
-  _A <- exists Type
+    => m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+    -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+    -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+app f a = do
+  _A <- meta type'
   x <- fresh
-  _B <- x ::: _A |- exists Type
-  let _F = pi (p :< Local (Name x) ::: case p of { Im -> zero ; Ex -> More } :@ _A) _B
+  _B <- ForAll x ::: _A |- meta type'
+  let _F = pi (Local x ::: _A) _B
   f' <- goalIs _F f
   a' <- goalIs _A a
-  pure (f' $$ (p :< a') ::: _F $$ (p :< a'))
+  pure (f' $$ a' ::: _F $$ a')
 
-
-exists :: ( Carrier sig m
-          , Member Naming sig
-          , Member (Reader (Context (Core (Name Meta)))) sig
-          , Member (State Signature) sig
-          )
-       => Core (Name Meta)
-       -> m (Core (Name Meta))
-exists ty = do
-  ctx <- ask
-  n <- fresh
-  let f (n ::: t) = Ex :< Local (Name n) ::: More :@ t
-      ty' = pis (f <$> Context.unContext ctx) ty
-  modify (Signature . Map.insert n ty' . unSignature)
-  pure (pure (Local (Meta n)) $$* ((Ex :<) . pure . Local . Name <$> Context.vars (ctx :: Context (Core (Name Meta)))))
 
 goalIs :: ( Carrier sig m
           , Member Naming sig
-          , Member (Reader (Context (Core (Name Meta)))) sig
-          , Member (Reader Span) sig
-          , Member (State Signature) sig
-          , Member (Writer (Set.Set (Spanned (Constraint Core (Name Meta))))) sig
           )
-       => Core (Name Meta)
-       -> m (Core (Name Meta) ::: Core (Name Meta))
-       -> m (Core (Name Meta))
+       => Term (Problem :+: Core) (Name Gensym)
+       -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
+       -> m (Term (Problem :+: Core) (Name Gensym))
 goalIs ty2 m = do
   tm1 ::: ty1 <- m
-  tm2 <- exists ty2
-  tm2 <$ unify (tm1 ::: ty1 :===: tm2 ::: ty2)
+  tm2 <- meta (ty1 === ty2)
+  pure (tm1 === tm2)
 
-unify :: (Carrier sig m, Member (Reader (Context (Core (Name Meta)))) sig, Member (Reader Span) sig, Member (Writer (Set.Set (Spanned (Constraint Core (Name Meta))))) sig)
-      => Equation (Core (Name Meta) ::: Core (Name Meta))
-      -> m ()
-unify (tm1 ::: ty1 :===: tm2 ::: ty2) = do
-  span <- ask
-  context <- asks (fmap (first (Local . Name)) . unContext)
-  tell (Set.fromList
-    [ binds context ((ty1 :===: ty2) ::: Type) :~ span
-    , binds context ((tm1 :===: tm2) ::: ty1)  :~ span
-    ])
+meta :: (Carrier sig m, Member Naming sig) => Term (Problem :+: Core) (Name Gensym) -> m (Term (Problem :+: Core) (Name Gensym))
+meta ty = do
+  n <- fresh
+  pure (exists (Local n ::: ty) (pure (Local n)))
 
-(|-) :: (Carrier sig m, Member (Reader (Context (Core (Name Meta)))) sig) => Gensym ::: Core (Name Meta) -> m a -> m a
-b |- m = local (Context.insert b) m
+(|-) :: (Carrier sig m, Member (Reader Context) sig) => Binding ::: Term (Problem :+: Core) (Name Gensym) -> m a -> m a
+b |- m = local (:> b) m
 
-infix 5 |-
+infix 3 |-
 
 
 elab :: ( Carrier sig m
         , Member (Error Doc) sig
         , Member Naming sig
-        , Member (Reader (Context (Core (Name Meta)))) sig
-        , Member (Reader (ModuleGraph Core Void)) sig
+        , Member (Reader Context) sig
         , Member (Reader Span) sig
-        , Member (State Signature) sig
-        , Member (Writer (Set.Set (Spanned (Constraint Core (Name Meta))))) sig
         )
      => Surface.Surface Qualified
-     -> m (Core (Name Meta) ::: Core (Name Meta))
+     -> m (Term (Problem :+: Core) (Name Gensym) ::: Term (Problem :+: Core) (Name Gensym))
 elab = Surface.kcata id alg bound assume
-  where bound (Z _) = asks @(Context (Core (Name Meta))) (first (pure . Local . Name) . Stack.head . unContext)
-        bound (S m) = local @(Context (Core (Name Meta))) (Context . Stack.tail . unContext) m
+  where bound (Z _) = asks @Context (first (Var . bindingName) . Stack.head)
+        bound (S m) = local @Context (Stack.drop 1) m
         alg = \case
-          Surface.Lam n b -> intro (unIgnored <$> n) (elab' (unScope <$> b))
-          (f Surface.:$ (p :< a)) -> app (elab' f) (p :< elab' a)
-          Surface.Type -> pure (Type ::: Type)
-          Surface.Pi (p :< Ignored n ::: t) b -> (p :< (n, elab' <$> t)) --> elab' (unScope <$> b)
+          Surface.Lam _ b -> intro (elab' (unScope <$> b))
+          f Surface.:$ (_ :< a) -> app (elab' f) (elab' a)
+          Surface.Type -> pure (type' ::: type')
+          Surface.Pi (_ :< _ ::: _ :@ t) b -> elab' t --> elab' (unScope <$> b)
         elab' = spanIs . fmap getConst
 
-type ElabC m = ReaderC (Context (Core (Name Meta))) (WriterC (Set.Set (Spanned (Constraint Core (Name Meta)))) m)
-
-runElab :: ElabC m a -> m (Set.Set (Spanned (Constraint Core (Name Meta))), a)
-runElab = runWriter . runReader mempty
-
-
-inferType :: (Carrier sig m, Member Naming sig) => m (Core (Name Meta))
-inferType = pure . Local . Meta <$> fresh
-
+elabDecl :: ( Carrier sig m
+            , Member (Error Doc) sig
+            , Member Naming sig
+            , Member (Reader Context) sig
+            , Member (Reader ModuleName) sig
+            )
+         => Decl (Surface.Surface Qualified)
+         -> m (Decl (Term (Problem :+: Core) Qualified))
+elabDecl (Decl name d tm ty) = namespace (show name) $ do
+  ty' <- runSpanned (goalIs type' . elab) ty
+  def <- meta (unSpanned ty')
+  moduleName <- ask
+  tm' <- runSpanned (local (:> Define (moduleName :.: name := def) ::: unSpanned ty') . goalIs (unSpanned ty') . elab) tm
+  ty'' <- runSpanned (either freeVariables pure . strengthen) ty'
+  tm'' <- runSpanned (either freeVariables pure . strengthen) tm'
+  pure (Decl name d tm'' ty'')
 
 elabModule :: ( Carrier sig m
-              , Effect sig
               , Member (Error Doc) sig
               , Member Naming sig
-              , Member (Reader (ModuleGraph Core Void)) sig
+              , Member (Reader (ModuleGraph (Term (Problem :+: Core)) Void)) sig
               , Member (Writer (Stack Doc)) sig
               )
            => Module Surface.Surface Qualified
-           -> m (Module Core Qualified)
-elabModule m = namespace (show (moduleName m)) . runReader (moduleName m) . local @(ModuleGraph Core Void) (Module.restrict (Map.keysSet (moduleImports m))) $ do
+           -> m (Module (Term (Problem :+: Core)) Qualified)
+elabModule m = namespace (show (moduleName m)) . runReader (moduleName m) . local @(ModuleGraph (Term (Problem :+: Core)) Void) (Module.restrict (Map.keysSet (moduleImports m))) $ do
   -- FIXME: do a topo sort on the decls? or at least make their types known first? or…?
   decls <- foldM go mempty (moduleDecls m)
   pure m { moduleDecls = decls }
-  where go decls decl = local (extendGraph decls) $ do
+  where go decls decl = local (extendGraph decls) . inContext $ do
           (extendModule decls <$> elabDecl (instantiate (pure . qualified . (moduleDecls m Map.!)) <$> decl)) `catchError` ((decls <$) . logError)
         extendModule decls decl = Map.insert (declName decl) (bind (Just . unqualified) <$> decl) decls
-        extendGraph decls (ModuleGraph g) = ModuleGraph @Core @Void (Map.insert (moduleName m) (bindHEither Left m { moduleDecls = decls }) g)
+        extendGraph decls (ModuleGraph g) = ModuleGraph @(Term (Problem :+: Core)) @Void (Map.insert (moduleName m) (bindHEither Left m { moduleDecls = decls }) g)
         qualified = (moduleName m :.:) . declName
         unqualified (_ :.: u) = u
 
+inContext :: (Carrier sig m, Member (Reader (ModuleGraph (Term (Problem :+: Core)) Void)) sig)
+          => ReaderC Context m a
+          -> m a
+inContext m = do
+  ctx <- asks @(ModuleGraph (Term (Problem :+: Core)) Void) toContext
+  runReader @Context ctx m
+  where toContext g = foldl' definitions Nil (modules g)
+        definitions ctx m = foldl' define ctx (moduleDecls m)
+          where define ctx d = ctx :> (Define ((moduleName m :.: declName d) := inst (declTerm d)) ::: inst (declType d))
+                inst t = instantiateEither (pure . Global . either (moduleName m :.:) id) (unSpanned t)
+
 logError :: (Member (Writer (Stack Doc)) sig, Carrier sig m) => Doc -> m ()
 logError = tell . (Nil :> )
-
-
-elabDecl :: ( Carrier sig m
-            , Effect sig
-            , Member (Error Doc) sig
-            , Member Naming sig
-            , Member (Reader (ModuleGraph Core Void)) sig
-            )
-         => Decl (Surface.Surface Qualified)
-         -> m (Decl (Core Qualified))
-elabDecl (Decl name d tm ty) = namespace (show name) $ do
-  ty' <- runSpanned (declare . elab) ty
-  (tm' ::: _) :~ s <- runSpanned (define (weaken (unSpanned ty')) . elab) tm
-  pure (Decl name d (tm' :~ s) ty')
-
-declare :: ( Carrier sig m
-           , Effect sig
-           , Member (Error Doc) sig
-           , Member Naming sig
-           , Member (Reader (ModuleGraph Core Void)) sig
-           , Member (Reader Span) sig
-           )
-        => ElabC (StateC Signature m) (Core (Name Meta) ::: Core (Name Meta))
-        -> m (Core Qualified)
-declare ty = evalState (mempty :: Signature) $ do
-  (constraints, ty') <- runElab (goalIs Type ty)
-  subst <- solver constraints
-  pure (generalizeType (apply subst ty'))
-
-define :: ( Carrier sig m
-          , Effect sig
-          , Member (Error Doc) sig
-          , Member Naming sig
-          , Member (Reader (ModuleGraph Core Void)) sig
-          , Member (Reader Span) sig
-          )
-       => Core (Name Meta)
-       -> ElabC (StateC Signature m) (Core (Name Meta) ::: Core (Name Meta))
-       -> m (Core Qualified ::: Core Qualified)
-define ty tm = evalState (mempty :: Signature) $ do
-  (constraints, tm') <- runElab (goalIs ty tm)
-  subst <- solver constraints
-  let tm'' = apply subst tm'
-      ty'  = generalizeType (apply subst ty)
-  (::: ty') <$> either (unsolvedMetavariables tm'') pure (strengthen tm'')
